Switch to an extension of Hanami::View itself
