+++ lib/hanami/configuration.rb
@@ -4,6 +4,7 @@ require "uri"
 require "concurrent/hash"
 require "concurrent/array"
 require "dry/inflector"
+require "pathname"
 
 module Hanami
   # Hanami application configuration
@@ -73,7 +74,7 @@ module Hanami
     end
 
     def root=(root)
-      settings[:root] = root
+      settings[:root] = Pathname(root)
     end
 
     def root
